♻️ refactor(tool_calls): relocate output handling functions
- move generate_id and convert_tool_calls_to_openai_format functions from input_handle.py to output_handle.py
- ensure consistent import structure across tool_calls modules
